+++ activerecord/lib/active_record/relation/predicate_builder/polymorphic_set_handler.rb
@@ -1,0 +1,6 @@
+module ActiveRecord
+  class PredicateBuilder
+    class PolymorphicSetHandler < PolymorphicArrayHandler; end # :nodoc:
+    class PolymorphicSetValue < PolymorphicArrayValue; end
+  end
+end